+++ android/src/main/java/com/tekartik/sqflite/operation/BaseReadOperation.java
@@ -10,7 +10,7 @@ import static com.tekartik.sqflite.Constant.PARAM_SQL_ARGUMENTS;
  * Created by alex on 09/01/18.
  */
 
-public abstract class BaseOperation extends BaseReadOperation {
+public abstract class BaseReadOperation implements Operation {
     public String getSql() {
         return getArgument(PARAM_SQL);
     }
@@ -26,15 +26,5 @@ public abstract class BaseOperation extends BaseReadOperation {
 
     // We actually have an inner object that does the implementation
     protected abstract OperationResult getResult();
-
-    @Override
-    public void success(Object results) {
-        getResult().success(results);
-    }
-
-    @Override
-    public void error(String errorCode, String errorMessage, Object data) {
-        getResult().error(errorCode, errorMessage, data);
-    }
 
 }